+++ lib/cloudflare_turnstile.dart
@@ -3,3 +3,4 @@ library cloudflare_turnstile;
 export 'src/widget/turnstile_options.dart';
 export 'src/controller/controller.dart';
 export 'src/widget/turnstile.dart';
+export 'src/widget/turnstile_form_field.dart';